minor code refactor

+++ Genetics/RealGene.cs
@@ -4,14 +4,14 @@ using Data;
 
 public class RealGene: Gene
 {
-    public double Value { get; }
+    public float Value { get; }
 
     public RealGene()
     {
-        this.Value = Gene.Random.NextDouble() * 2 - 1;
+        this.Value = (float) (Gene.Random.NextDouble() * 2 - 1);
     }
 
-    private RealGene(double value)
+    private RealGene(float value)
     {
         this.Value = value;
     }
@@ -23,7 +23,7 @@ public class RealGene: Gene
 
     public override Gene Mutated(ApproximatorJob job)
     {
-        var delta = 1.0 / Math.Pow(10, job.PrecisionDigits);
+        var delta = 1.0F / (float) Math.Pow(10, job.PrecisionDigits);
         delta *= Gene.Random.Next() % 2 == 1 ? -1 : 1;
         return new RealGene(this.Value + delta);
     }
